Move feature_output to kart/tabular
